Split user_agent() to it's own module.
So 'download' and 'rest_api' modules can use it without
circular deps.

Update rest_api to use user_agent module
