+++ ansible_galaxy/user_agent.py
@@ -1,0 +1,16 @@
+import sys
+
+from ansible_galaxy import __version__ as mazer_version
+
+USER_AGENT_FORMAT = 'Mazer/{version} ({platform}; python:{py_major}.{py_minor}.{py_micro}) ansible_galaxy/{version}'
+
+
+def user_agent():
+    '''Return a user-agent including mazer version, platform, and python version'''
+
+    user_agent_data = {'version': mazer_version,
+                       'platform': sys.platform,
+                       'py_major': sys.version_info.major,
+                       'py_minor': sys.version_info.minor,
+                       'py_micro': sys.version_info.micro}
+    return USER_AGENT_FORMAT.format(**user_agent_data)
